airbnb_poetry/categories/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from .serializers import CategorySerializer
from .models import Category


class Categories(APIView):
    def get(self, request):
        categories = Category.objects.all()
        serializer = CategorySerializer(categories, many=True)
        return Response(serializer.data)


class CategoryDetail(APIView):
    def get(self, request, pk):
        category = Category.objects.get(pk=pk)
        serializer = CategorySerializer(category)
        return Response(serializer.data)
